+++ src/command.ts
@@ -1,8 +1,8 @@
 // insert new commands here
-import ping from './commands/ping'
-import addproject from './commands/addproject'
+import ping from './interactions/chat-input-command/ping'
+import addproject from './interactions/chat-input-command/addproject'
 import { Collection } from 'discord.js'
-import edit from './commands/edit'
+import edit from './interactions/context-menu-command/edit-project'
 
 // enable this to deploy commands
 const DEPLOY_COMMANDS = false
