+++ src/interactions/chat-input-command/ping.ts
